* bld/warn:
  cosmetics, use proper format specifier for lzma_ret values
  cosmetics, shut icc up about shortening ints in uta printer
  cosmetics, shut icc up about shortening ints in ute-info

+++ src/uta.c
@@ -236,8 +236,8 @@ read_line(mux_ctx_t ctx, struct sndwch_s *tl)
 	char *line;
 	/* symbol and its index */
 	const char *sym;
-	unsigned int symidx;
-	unsigned int ttf;
+	uint16_t symidx;
+	uint16_t ttf;
 
 	/* get the line, its length and set up the cursor */
 	if (UNLIKELY(prchunk_getline(ctx->rdr, &line) <= 0)) {
@@ -258,7 +258,7 @@ read_line(mux_ctx_t ctx, struct sndwch_s *tl)
 
 	/* next up is the sym-idx in hex */
 	if (LIKELY(*cursor != '?')) {
-		symidx = hex2int(&cursor);
+		symidx = (uint16_t)hex2int(&cursor);
 		if (UNLIKELY(*cursor++ != '\t')) {
 			return -1;
 		}
@@ -280,14 +280,14 @@ read_line(mux_ctx_t ctx, struct sndwch_s *tl)
 	}
 
 	/* check the tick type + flags, it's hex already */
-	ttf = hex2int(&cursor);
+	ttf = (uint16_t)hex2int(&cursor);
 	if (UNLIKELY(*cursor++ != '\t')) {
 		return -1;
 	}
 
 	/* bang it all into the target tick */
-	scom_thdr_set_tblidx(AS_SCOM_THDR(tl), (uint16_t)symidx);
-	scom_thdr_set_ttf(AS_SCOM_THDR(tl), (uint16_t)ttf);
+	scom_thdr_set_tblidx(AS_SCOM_THDR(tl), symidx);
+	scom_thdr_set_ttf(AS_SCOM_THDR(tl), ttf);
 
 	/* now on to the payload */
 	if (ttf < SCOM_FLAG_LM && UNLIKELY(__u64_payload_p(ttf))) {
